New folder/requirements.txt
# Core dependencies
opencv-python>=4.5.0
numpy>=1.19.0
easyocr>=1.6.0
Pillow>=8.0.0
PyJWT>=2.0.0

# Web interface
flask>=2.0.0
flask-socketio>=5.0.0
eventlet>=0.30.0

# Utilities
python-dateutil>=2.8.0
requests>=2.25.0